added: OnlyofficeBaseFormatter
